Handle Citations on Project and Footnotes on Document Level
