+++ packages/guides/src/Compiler/NodeTransformers/CitationInlineNodeTransformer.php
@@ -6,22 +6,16 @@ namespace phpDocumentor\Guides\Compiler\NodeTransformers;
 
 use phpDocumentor\Guides\Compiler\CompilerContext;
 use phpDocumentor\Guides\Compiler\NodeTransformer;
-use phpDocumentor\Guides\Metas;
 use phpDocumentor\Guides\Nodes\Inline\CitationInlineNode;
 use phpDocumentor\Guides\Nodes\Node;
 
 /** @implements NodeTransformer<Node> */
 class CitationInlineNodeTransformer implements NodeTransformer
 {
-    public function __construct(
-        private readonly Metas $metas,
-    ) {
-    }
-
     public function enterNode(Node $node, CompilerContext $compilerContext): Node
     {
         if ($node instanceof CitationInlineNode) {
-            $internalTarget = $this->metas->getCitationTarget($node->getName());
+            $internalTarget = $compilerContext->getProjectNode()->getCitationTarget($node->getName());
             $node->setInternalTarget($internalTarget);
         }
