+++ packages/guides/src/Compiler/NodeTransformers/FootNodeNamedTransformer.php
@@ -7,22 +7,16 @@ namespace phpDocumentor\Guides\Compiler\NodeTransformers;
 use phpDocumentor\Guides\Compiler\CompilerContext;
 use phpDocumentor\Guides\Compiler\NodeTransformer;
 use phpDocumentor\Guides\Meta\FootnoteTarget;
-use phpDocumentor\Guides\Metas;
 use phpDocumentor\Guides\Nodes\FootnoteNode;
 use phpDocumentor\Guides\Nodes\Node;
 
 /** @implements NodeTransformer<Node> */
 class FootNodeNamedTransformer implements NodeTransformer
 {
-    public function __construct(
-        private readonly Metas $metas,
-    ) {
-    }
-
     public function enterNode(Node $node, CompilerContext $compilerContext): Node
     {
         if ($node instanceof FootnoteNode && $this->supports($node)) {
-            $number = $this->metas->addFootnoteTarget(new FootnoteTarget(
+            $number = $compilerContext->getDocumentNode()->addFootnoteTarget(new FootnoteTarget(
                 $compilerContext->getDocumentNode()->getFilePath(),
                 $node->getAnchor(),
                 $node->getName(),
