+++ tests/Integration/tests/footnotes-on-document/expected/index.html
@@ -1,0 +1,27 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <title>Document Title</title>
+
+</head>
+<body>
+    <div class="section" id="document-title">
+    <h1>Document Title</h1>
+
+            <p>Lorem ipsum <sup>[<a href="index.html#footnote-1">1</a>]</sup> dolor sit amet <sup>[<a href="index.html#footnote-2">2</a>]</sup>.</p>
+            <div class="rubric">Footnotes</div>
+
+            <div class="footnote " id="footnote-1">
+    <div class="footnote-label">[1]</div>
+    <div class="footnote-content">Some numbered footnote</div>
+</div>
+
+            <div class="footnote " id="footnote-2">
+    <div class="footnote-label">[2]</div>
+    <div class="footnote-content">Another numbered footnote</div>
+</div>
+
+    </div>
+
+</body>
+</html>
