+++ tests/Integration/tests/footnotes-on-document/expected/page1.html
@@ -1,0 +1,27 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <title>Page 1</title>
+
+</head>
+<body>
+    <div class="section" id="page-1">
+    <h1>Page 1</h1>
+
+            <p>Lorem ipsum <sup>[<a href="page1.html#footnote-1">1</a>]</sup> dolor sit amet <sup>[<a href="page1.html#footnote-2">2</a>]</sup>.</p>
+            <div class="rubric">Footnotes</div>
+
+            <div class="footnote " id="footnote-1">
+    <div class="footnote-label">[1]</div>
+    <div class="footnote-content">Some numbered footnote</div>
+</div>
+
+            <div class="footnote " id="footnote-2">
+    <div class="footnote-label">[2]</div>
+    <div class="footnote-content">Another numbered footnote</div>
+</div>
+
+    </div>
+
+</body>
+</html>
