+++ tests/Integration/tests/footnotes-on-document/input/index.rst
@@ -1,0 +1,11 @@
+==============
+Document Title
+==============
+
+Lorem ipsum [1]_ dolor sit amet [2]_.
+
+..  rubric:: Footnotes
+
+.. [1] Some numbered footnote
+
+.. [2] Another numbered footnote
